refactor(cli): move default (evaluate) command action to its own file
